requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
sqlalchemy>=2.0
pyarrow>=12.0
pillow>=10.0

